// apps/b2c/src/modules/profile/types/profile-data-copy.ts
export type ProfileDataCopy = {
  formTitle: string
  firstNameLabel: string
  firstNamePlaceholder: string
  lastNameLabel: string
  lastNamePlaceholder: string
  middleNameLabel: string
  middleNamePlaceholder: string
  phoneLabel: string
  phonePlaceholder: string
  emailLabel: string
  emailPlaceholder: string
  passportLabel: string
  passportPlaceholder: string
  birthDateLabel: string
  birthDatePlaceholder: string
  cancelLabel: string
  saveLabel: string
}

export type ProfileSidebarCopy = {
  customerData: string
  myBookings: string
  favorites: string
  logout: string
}